requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
gunicorn>=21.0.0
